managed-connectivity/sample-custom-connector/requirements.txt
google-cloud-dataplex==2.2.2
google-cloud-storage
google-cloud-secret-manager
